+++ tests/sht3x_i2c_test.cpp
@@ -3,7 +3,7 @@
  *
  * Generator:     sensirion-driver-generator 0.33.0
  * Product:       sht3x
- * Model-Version: 1.0.0
+ * Model-Version: 2.0.0
  */
 
 #include "sensirion_common.h"
@@ -21,17 +21,6 @@ TEST_GROUP (SHT3X_Tests) {
         sht3x_init(0x44);
     }
 };
-
-TEST (SHT3X_Tests, test_measure_single_shot1) {
-    int16_t local_error = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
-    local_error = sht3x_measure_single_shot(REPEATABILITY_MEDIUM, false,
-                                            &a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "measure_single_shot");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
-}
 
 TEST (SHT3X_Tests, test_measure_single_shot_high_repeatability1) {
     int16_t local_error = 0;
@@ -151,8 +140,6 @@ TEST (SHT3X_Tests, test_start_periodic_measurement1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_periodic_measurement(REPEATABILITY_MEDIUM,
                                                    MPS_ONE_PER_SECOND);
     CHECK_EQUAL_ZERO_TEXT(local_error, "start_periodic_measurement");
@@ -160,10 +147,6 @@ TEST (SHT3X_Tests, test_start_periodic_measurement1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -172,18 +155,12 @@ TEST (SHT3X_Tests, test_start_art_measurement1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_art_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "start_art_measurement");
     local_error = sht3x_read_measurement(&temperature_ticks, &humidity_ticks);
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -192,8 +169,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_high_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_0_5_mps_high_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_0_5_mps_high_repeatability");
@@ -201,10 +176,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_high_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -213,8 +184,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_medium_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_0_5_mps_medium_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_0_5_mps_medium_repeatability");
@@ -222,10 +191,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_medium_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -234,8 +199,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_low_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_0_5_mps_low_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_0_5_mps_low_repeatability");
@@ -243,10 +206,6 @@ TEST (SHT3X_Tests, test_start_measurement_0_5_mps_low_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -255,8 +214,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_high_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_1_mps_high_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_1_mps_high_repeatability");
@@ -264,10 +221,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_high_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -276,8 +229,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_medium_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_1_mps_medium_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_1_mps_medium_repeatability");
@@ -285,10 +236,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_medium_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -297,8 +244,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_low_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_1_mps_low_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_1_mps_low_repeatability");
@@ -306,10 +251,6 @@ TEST (SHT3X_Tests, test_start_measurement_1_mps_low_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -318,8 +259,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_high_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_2_mps_high_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_2_mps_high_repeatability");
@@ -327,10 +266,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_high_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -339,8 +274,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_medium_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_2_mps_medium_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_2_mps_medium_repeatability");
@@ -348,10 +281,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_medium_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -360,8 +289,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_low_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_2_mps_low_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_2_mps_low_repeatability");
@@ -369,10 +296,6 @@ TEST (SHT3X_Tests, test_start_measurement_2_mps_low_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -381,8 +304,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_high_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_4_mps_high_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_4_mps_high_repeatability");
@@ -390,10 +311,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_high_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -402,8 +319,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_medium_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_4_mps_medium_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_4_mps_medium_repeatability");
@@ -411,10 +326,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_medium_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -423,8 +334,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_low_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_4_mps_low_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_4_mps_low_repeatability");
@@ -432,10 +341,6 @@ TEST (SHT3X_Tests, test_start_measurement_4_mps_low_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -444,8 +349,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_high_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_10_mps_high_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_10_mps_high_repeatability");
@@ -453,10 +356,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_high_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -465,8 +364,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_medium_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_10_mps_medium_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_10_mps_medium_repeatability");
@@ -474,10 +371,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_medium_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }
@@ -486,8 +379,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_low_repeatability1) {
     int16_t local_error = 0;
     uint16_t temperature_ticks = 0;
     uint16_t humidity_ticks = 0;
-    float a_temperature = 0.0;
-    float a_humidity = 0.0;
     local_error = sht3x_start_measurement_10_mps_low_repeatability();
     CHECK_EQUAL_ZERO_TEXT(local_error,
                           "start_measurement_10_mps_low_repeatability");
@@ -495,10 +386,6 @@ TEST (SHT3X_Tests, test_start_measurement_10_mps_low_repeatability1) {
     CHECK_EQUAL_ZERO_TEXT(local_error, "read_measurement");
     printf("temperature_ticks: %u ", temperature_ticks);
     printf("humidity_ticks: %u\n", humidity_ticks);
-    local_error = sht3x_blocking_read_measurement(&a_temperature, &a_humidity);
-    CHECK_EQUAL_ZERO_TEXT(local_error, "blocking_read_measurement");
-    printf("a_temperature: %.2f ", a_temperature);
-    printf("a_humidity: %.2f\n", a_humidity);
     local_error = sht3x_stop_measurement();
     CHECK_EQUAL_ZERO_TEXT(local_error, "stop_measurement");
 }